Resolve most of remaining ErrorProne issues in :constraintlayout
Resolved Unused parameter, ShortCircuitBoolean,
NarrowingCompoundAssignment, Unused field.

Removed fields that had no use whatsoever (not referenced, not
assigned).

+++ constraintlayout/constraintlayout/src/main/java/androidx/constraintlayout/motion/widget/MotionController.java
@@ -484,7 +484,7 @@ public class MotionController {
             mSpline[0].getPos(p, mInterpolateData);
             mStartMotionPath.getCenter(p, mInterpolateVariables, mInterpolateData, points, 0);
             if (i > 0) {
-                sum += Math.hypot(y - points[1], x - points[0]);
+                sum += (float) Math.hypot(y - points[1], x - points[0]);
             }
             x = points[0];
             y = points[1];
